refactor: change path change the path to get a style
